Modify to fit all images ratio
Because of minimum height(min-height) setting, some images are out of ratio.
The image width and height settings are applied differently to keep the ratio.
Images with a width greater than the height set the height to 100% and the width to auto.
Other images (with a height greater than the width) set the width to 100% and the height to auto.